refactor: consolidate canvas protocol types
- group requests and responses together
- TG-253
