--- palette-types/src/canvasProtocol/CanvasAPIError.ts
@@ -1,6 +1,0 @@
-/**
- * When Canvas returns an error, this object will be in the error field of the response.
- */
-export interface CanvasAPIError {
-  message: string;
-}
