--- palette-types/src/canvasProtocol/CanvasAPIErrorResponse.ts
@@ -1,5 +1,0 @@
-import { CanvasAPIError } from "./CanvasAPIError";
-
-export interface CanvasAPIErrorResponse {
-  errors: CanvasAPIError[];
-}
