--- palette-types/src/canvasProtocol/CreateRubricAssociationResponse.ts
@@ -1,6 +1,0 @@
-import { CanvasAssociation } from "../canvasTypes/CanvasAssociation";
-import { CanvasAPIErrorResponse } from "./CanvasAPIErrorResponse";
-
-export type CreateRubricAssociationResponse =
-  | CanvasAssociation
-  | CanvasAPIErrorResponse;
