--- palette-types/src/canvasProtocol/DeleteRubricResponse.ts
@@ -1,4 +1,0 @@
-import { CanvasRubric } from "../canvasTypes/CanvasRubric";
-import { CanvasAPIErrorResponse } from "./CanvasAPIErrorResponse";
-
-export type DeleteRubricResponse = CanvasRubric | CanvasAPIErrorResponse;
