--- palette-types/src/canvasProtocol/GetAllRubricsRequest.ts
@@ -1,3 +1,0 @@
-export interface GetAllRubricsRequest {
-  courseID: number;
-}
