--- palette-types/src/canvasProtocol/GetRubricResponse.ts
@@ -1,4 +1,0 @@
-import { CanvasRubric } from "../canvasTypes/CanvasRubric";
-import { CanvasAPIErrorResponse } from "./CanvasAPIErrorResponse";
-
-export type GetRubricResponse = CanvasRubric | CanvasAPIErrorResponse;
